+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to
 [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.21.0] - 2018-02-18
+### Changed
+- `VehicleGroupStatistics` can now be sorted by count, make, or model, in
+  ascending or descending order.
+
 ## [1.20.0] - 2018-02-14
 ### Added
 - `./bin/deploy` shortcut for `yarn deploy`.